Remove test file from lib target. Link libleveldb into framework.
To save client applications from having to link to both the framework and the underlying leveldb static lib, link the latter into the former. This may require some finessing since the framework is supposed to be universal, but the leveldb lib target only builds for one platform at a time.