Include the classname in the array token.
We need to know whether we should dereference the value stored in the array index key one more time to get the object.

I'm not happy with this solution or the code layout; it's doing too much work and the object loading has to know about the array index key format.