Starting to define a C++ API for supporting object graph serialization.
We have a number of key/token types that follow a consistent pattern. Inevitably you have to start with an application-specific unique identifier for the root object and any other serializable objects. Similarly, information about the class must be stored under a unique string for the class, based on the class name.

At first I thought I might store the information about classes in a single dictionary stored under a private key, but this would not scale well. We have a dictionary interface in leveldb itself, so we should just use it. I just don't yet really understand how easy it will be to, for example, find all the class information using the Iterator interface (specifically, what constraints are placed on key patterns to make sure key search and iteration works properly).

An object's key recovers it's token. From the token you can retrieve the class number, with which you can generate the class key and recover the class description, which includes the object properties. Or since you're instantiating an object anyway, ask it for the property names. (In future we will need to support versioning, which is why we want to store the properties in the database). Property keys are constructed using the 

Array-type properties are stored as an array token, from which can be constructed array index keys.

There are some errors in the existing specification and related classes, so I'm committing what I have now before I work on repairing it.