Simplify key design and usage.
We don't need an object token or class token after all. We replace them with the object key and class key we designed previously, since we do need to use them as lookups. The object key lets us look up the object name, and the class key lets us look up the class description.

We also use the object key to create a class key which lets us look up the class description, from which we get the class name, instantiate an object, ask for its properties, then generate property keys and use them to find the property values.