Add tests for the size checking functionality recently added.
The tests fail because of some unspecified requirements of the underlying leveldb code. The GetApproximateSizes() method depends on Version.ApproximateOffsetOf(), which depends on the internal key comparator which doesn't support recognizing when two keys are the same. So the size check doesn't find the given key, at least when there are too few keys. I don't know how many keys are enough, because it fails for 32 keys.