Rewrite the basic operation tests in the new style.
This is a rather complex design, meant to avoid a lot of copy and paste coding. The test consists of a wrapper and additional data in the form of operation blocks. The test set data is prepared with different blocks which perform stores, loads and deletes specific to the kind of data being used (and thus the interface being tested). There is still quite a bit of repetition in the code, but not in the logic, only in preamble.