Correction to build settings fix.
The previous changed did not fix the build of leveldb, which was then linked into NULevelDB without all architectures. A cursory examination with lipo would seem to show that all three architectures were present, but the NULevelDB armv6 architecture was not accompanied by binaries for leveldb classes in armv6. I don't understand how that build could be called successful, when the armv6 builds of the LevelDB classes had unresolved links to leveldb symbols!